requirements.txt
black==22.1.0
click==8.0.3
install==1.3.5
mypy-extensions==0.4.3
mysql-connector-python==8.0.28
pathspec==0.9.0
platformdirs==2.5.0
protobuf==3.19.4
tomli==2.0.1
typing-extensions==4.1.1
